+++ tests/test_properties.py
@@ -2,7 +2,17 @@ import numpy as np
 import pint
 from pytest import approx
 
-from spice_segmenter.trajectory_properties import Distance, PhaseAngle, UnitAdaptor
+from spice_segmenter.decorators import declare
+from spice_segmenter.ops import Inverted
+from spice_segmenter.trajectory_properties import (
+    Constant,
+    Constraint,
+    Distance,
+    PhaseAngle,
+    Property,
+    UnitAdaptor,
+)
+from spice_segmenter.types import TIMES_TYPES
 
 from . import TourConfig
 
@@ -10,6 +20,11 @@ tc = TourConfig()
 start, end = tc.coverage
 
 t1 = start + np.timedelta64(100, "D")
+
+
+def test_imports() -> None:
+    # some random imports
+    pass
 
 
 def test_distance() -> None:
@@ -53,3 +68,34 @@ def test_unit_adaptor() -> None:
     ph_deg = ph_deg(t1)
 
     assert ph_deg == approx(np.rad2deg(ph(t1)))
+
+
+# property declare is a wrapper that can automatically create a new Property class from a function and some more arguments.
+# it is used to create properties from functions that are not properties themselves.
+
+
+@declare(name="one_property")
+class FakeOneProperty(Property):
+    def __call__(self, time: TIMES_TYPES) -> float:
+        return 1
+
+
+def test_inverted_wrapped():
+    p = FakeOneProperty()
+    value = p("2020-01-01")
+    assert value == 1
+
+    c = Constraint(p, 1, "==")
+    assert isinstance(
+        c.right, Constant
+    )  # it should translate a 1 to a constant by itself.
+
+    c2 = p == 1
+    assert isinstance(
+        c2.right, Constant
+    )  # it should translate a 1 to a constant by itself.
+
+    assert c("2020-01-01") == True
+    p_ = Inverted(c)
+    value = p_("2020-01-01")
+    assert value == False